Fix to work with renamed classes
