+++ ui-sdk-components-sample/app/src/main/java/com/karhoo/samples/uisdk/components/booking/TripBookingFragment.kt
@@ -11,15 +11,15 @@ import com.karhoo.samples.uisdk.components.R
 import com.karhoo.samples.uisdk.components.base.BaseFragment
 import com.karhoo.sdk.api.model.Quote
 import com.karhoo.uisdk.screen.booking.booking.payment.adyen.AdyenPaymentView
-import com.karhoo.uisdk.screen.booking.booking.supplier.BookingSupplierViewModel
-import com.karhoo.uisdk.screen.booking.booking.supplier.QuoteListStatus
+import com.karhoo.uisdk.screen.booking.booking.quotes.BookingQuotesViewModel
+import com.karhoo.uisdk.screen.booking.booking.quotes.QuoteListStatus
 import com.karhoo.uisdk.screen.booking.domain.address.BookingStatusStateViewModel
 import com.karhoo.uisdk.screen.booking.domain.bookingrequest.BookingRequestStateViewModel
 import kotlinx.android.synthetic.main.fragment_trip_booking.*
 
 class TripBookingFragment : BaseFragment() {
 
-    private lateinit var bookingSupplierViewModel: BookingSupplierViewModel
+    private lateinit var bookingQuotesViewModel: BookingQuotesViewModel
     private lateinit var bookingStatusStateViewModel: BookingStatusStateViewModel
     private lateinit var bookingRequestStateViewModel: BookingRequestStateViewModel
     private var quote: Quote? = null
@@ -72,13 +72,13 @@ class TripBookingFragment : BaseFragment() {
         fun newInstance(
             owner: LifecycleOwner,
             bookingStatusStateViewModel: BookingStatusStateViewModel,
-            bookingSupplierViewModel: BookingSupplierViewModel,
+            bookingQuotesViewModel: BookingQuotesViewModel,
             bookingRequestStateViewModel: BookingRequestStateViewModel
         ) = TripBookingFragment().apply {
             this.bookingStatusStateViewModel = bookingStatusStateViewModel
-            this.bookingSupplierViewModel = bookingSupplierViewModel
+            this.bookingQuotesViewModel = bookingQuotesViewModel
             this.bookingRequestStateViewModel = bookingRequestStateViewModel
-            bookingSupplierViewModel.viewStates().observe(owner, createQuoteObservable())
+            bookingQuotesViewModel.viewStates().observe(owner, createQuoteObservable())
         }
     }
 }
