+++ ui-sdk-components-sample/app/src/main/res/layout/fragment_trip_quotes.xml
@@ -8,8 +8,8 @@
     android:padding="@dimen/padding_default"
     tools:context="com.karhoo.samples.uisdk.components.quotes.TripQuotesFragment">
 
-    <com.karhoo.uisdk.screen.booking.supplier.SupplierListView
-        android:id="@+id/supplier_list_widget"
+    <com.karhoo.uisdk.screen.booking.quotes.QuotesListView
+        android:id="@+id/quotes_list_widget"
         android:layout_width="0dp"
         android:layout_height="@dimen/collapsible_pane_expanded_height"
         app:layout_constraintEnd_toEndOf="parent"
